+++ R/modelsummary.R
@@ -11,7 +11,7 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low', 'val
 #' @param output filename or object type (character string)
 #' \itemize{
 #'   \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.
-#'   \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
+#'   \item Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
 #'   \item Warning: the `output` argument \emph{cannot} be used when customizing tables with external packages. See the 'Details' section below.
 #' }
 #' @param fmt determines how to format numeric values
@@ -75,7 +75,7 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low', 'val
 #' the data.frame produced by `get_estimates(model)`. Examples:
 #' \itemize{
 #'   \item "estimate"
-#'   \item "\{estimate\} (\{std.error\}){stars}"
+#'   \item "\{estimate\} (\{std.error\})\{stars\}"
 #'   \item "\{estimate\} [\{conf.low\}, \{conf.high\}]"
 #' }
 #' @param align A character string of length equal to the number of columns in