+++ NEWS.md
@@ -3,7 +3,7 @@
 * new output format: latex_tabular
 * `tidy_custom` allows partial term matches
 * modelsummary(coef_rename=) accepts functions
-* new function `coef_rename` for use in modelsummary(coef_rename=)
+* new function `coef_rename` for use in modelsummary(coef_rename=coef_rename)
 * `modelplot` accepts `add_rows` to add reference categories
 * bug fixes